+++ AchtungWebsocket/app/models/iface/Collidable.java
@@ -1,6 +1,5 @@
 package models.iface;
 
 public interface Collidable {
-	public boolean isCollision(Long x, Long y);
-
+	public boolean isCollision(Double x, Double y);
 }
